Change backend record model to store time as float

+++ Typing App API/Models/Record.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public int Time { get; set; }
+        public float Time { get; set; }
         public User? User { get; set; }
     }
 }
